Add a context-menu item to change a team's side on player's turn
Essentially, this allows the player to, on-the-fly, consider another
team as an ally or an enemy during their turn.
At the beginning of each turn, each team's team_name is cached (currently
including the player's team, as Lua objects to `wml.tag.not` when trying
to use `wesnoth.sides.find`.
Then on the player's turn, each side's team_name is set according to
shownTeams (eg shownTeams as `{2="EnemyGuys", 3="Indrith"}). After the
player's turn, their team names are set back to their real ones.

...Note to self. There may be an edge case here. If a team's allegiance
is Actually Changed during the player's turn, then at the end of the
player's turn, we want to check if team names are different to what we
set them to. If they are, then they've been changed during, and we want
to add that changed team name to the cache before resetting.